mgsrc/mcp_bridge/requirements.txt
python-socketio[client]>=5.11.0
pydantic>=2.6.0

